requirements.txt
Flask==3.1.0
mysql-connector-python==9.0.0
python-dotenv==1.0.1